+++ plugins/woocommerce-admin/src/API/ProductAttributeTerms.php
@@ -15,6 +15,8 @@ defined( 'ABSPATH' ) || exit;
  * @extends WC_REST_Product_Attribute_Terms_Controller
  */
 class ProductAttributeTerms extends \WC_REST_Product_Attribute_Terms_Controller {
+	use CustomAttributeTraits;
+
 	/**
 	 * Endpoint namespace.
 	 *
@@ -96,29 +98,51 @@ class ProductAttributeTerms extends \WC_REST_Product_Attribute_Terms_Controller 
 			return array();
 		}
 
+		$attribute_values = array();
+
+		// Get the attribute properties.
+		$attribute = $this->get_custom_attribute_by_slug( $slug );
+
+		if ( is_wp_error( $attribute ) ) {
+			return $attribute;
+		}
+
 		// Find all attribute values assigned to products.
 		$query_results = $wpdb->get_results(
 			$wpdb->prepare(
 				"SELECT meta_value, COUNT(meta_id) AS product_count
 				FROM {$wpdb->postmeta}
 				WHERE meta_key = %s
+				AND meta_value != ''
 				GROUP BY meta_value",
 				'attribute_' . esc_sql( $slug )
 			),
-			ARRAY_A
+			OBJECT_K
 		);
 
-		$attribute_values = array();
+		// Ensure all defined properties are in the response.
+		$defined_values = wc_get_text_attributes( $attribute[ $slug ]['value'] );
 
-		foreach ( $query_results as $term ) {
+		foreach ( $defined_values as $defined_value ) {
+			if ( array_key_exists( $defined_value, $query_results ) ) {
+				continue;
+			}
+
+			$query_results[ $defined_value ] = (object) array(
+				'meta_value'    => $defined_value, // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_value
+				'product_count' => 0,
+			);
+		}
+
+		foreach ( $query_results as $term_value => $term ) {
 			// Mimic the structure of a taxonomy-backed attribute values for response.
 			$data = array(
-				'id'          => $term['meta_value'],
-				'name'        => $term['meta_value'],
-				'slug'        => $term['meta_value'],
+				'id'          => $term_value,
+				'name'        => $term_value,
+				'slug'        => $term_value,
 				'description' => '',
 				'menu_order'  => 0,
-				'count'       => (int) $term['product_count'],
+				'count'       => (int) $term->product_count,
 			);
 
 			$response = rest_ensure_response( $data );
@@ -133,10 +157,10 @@ class ProductAttributeTerms extends \WC_REST_Product_Attribute_Terms_Controller 
 			);
 			$response = $this->prepare_response_for_collection( $response );
 
-			$attribute_values[] = $response;
+			$attribute_values[ $term_value ] = $response;
 		}
 
-		return $attribute_values;
+		return array_values( $attribute_values );
 	}
 
 	/**