+++ plugins/woocommerce-admin/src/API/Reports/Orders/DataStore.php
@@ -65,7 +65,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 		$table_name = self::get_db_table_name();
 		// Avoid ambigious columns in SQL query.
 		$this->report_columns = array(
-			'order_id'         => "{$table_name}.order_id",
+			'order_id'         => "DISTINCT {$table_name}.order_id",
 			'parent_id'        => "{$table_name}.parent_id",
 			'date_created'     => "{$table_name}.date_created",
 			'date_created_gmt' => "{$table_name}.date_created_gmt",
@@ -181,7 +181,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 			$where_subquery[] = "{$order_tax_lookup_table}.tax_rate_id NOT IN ({$excluded_tax_rates}) OR {$order_tax_lookup_table}.tax_rate_id IS NULL";
 		}
 
-		$attribute_subqueries = $this->get_attribute_subqueries( $query_args );
+		$attribute_subqueries = $this->get_attribute_subqueries( $query_args, $order_stats_lookup_table );
 		if ( $attribute_subqueries['join'] && $attribute_subqueries['where'] ) {
 			$this->subquery->add_sql_clause( 'join', "JOIN {$order_product_lookup_table} ON {$order_stats_lookup_table}.order_id = {$order_product_lookup_table}.order_id" );