+++ plugins/woocommerce-admin/tests/api/product-attributes.php
@@ -51,9 +51,34 @@ class WC_Tests_API_Product_Attributes extends WC_REST_Unit_Test_Case {
 		$product->set_attributes( $attributes );
 		$product->save();
 
-		// Custom attribute terms can only be found once assigned to variations.
-		$data_store = $product->get_data_store();
-		$data_store->create_all_product_variations( $product );
+		// Assign one variation to the '1' size.
+		$variation  = $product->get_available_variations( 'objects' )[0];
+		$attributes = $variation->get_attributes();
+		$attributes[ sanitize_title( $custom_attr->get_name() ) ] = '1';
+		$variation->set_attributes( $attributes );
+		$variation->save();
+
+		// Add more custom Numeric Size values to another product.
+		$product_2 = new WC_Product_Variable();
+		$product_2->set_props(
+			array(
+				'name' => 'Dummy Variable Product 2',
+				'sku'  => 'DUMMY VARIABLE SKU 2',
+			)
+		);
+
+		$custom_attr_2 = new WC_Product_Attribute();
+		$custom_attr_2->set_name( 'Numeric Size' );
+		$custom_attr_2->set_options( array( '6', '7', '8', '9', '10' ) );
+		$custom_attr_2->set_visible( true );
+		$custom_attr_2->set_variation( true );
+
+		$product_2->set_attributes(
+			array(
+				$custom_attr_2,
+			)
+		);
+		$product_2->save();
 	}
 
 	/**
@@ -228,10 +253,9 @@ class WC_Tests_API_Product_Attributes extends WC_REST_Unit_Test_Case {
 		$terms    = $response->get_data();
 
 		$this->assertEquals( 200, $response->get_status() );
-		$this->assertEquals( 5, count( $terms ) );
-
-		// All terms will have the same count since we created all possible variations.
-		// Test for ( 3 size * 2 colour * 3 number ) combinations = 18.
-		$this->assertEquals( 18, $terms[0]['count'] );
+		$this->assertEquals( 10, count( $terms ) );
+		$this->assertEquals( '1', $terms[0]['slug'] );
+		$this->assertEquals( 1, $terms[0]['count'] );
+		$this->assertEquals( 0, $terms[1]['count'] );
 	}
 }