+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -15,6 +15,7 @@
 -   Deprecate the Count component, with plan to remove in next major version. #7115
 -   Remove the long deprecated Card component (use Card from `@wordpress/components` instead). #7114
 -   Add `<AbbreviatedCard />` component. #7017
+-   Fix non-string query prop warning in SelectControl component. #7046
 -   Remove support for IE11. #7112
 
 # 6.2.0